refactored login as can component
