+++ assets/js/components/login-component/demo.html
@@ -1,0 +1,17 @@
+<!doctype html>
+
+<script type="text/stache" id="demo" can-autorender>
+  <can-import from="components/login-component/" />
+  <can-import from="bootstrap" />
+  <can-import from="bootstrap.css!" />
+  <login-component {user-name}="userName" is-shown="true"></login-component>
+</script>
+
+<script src="../../../../node_modules/steal/steal.js" main="can/view/autorender/">
+  import can from 'can';
+  import $ from 'jquery';
+
+  $('#demo').viewModel().attr({
+    userName: 'Petya'
+  });
+</script>
