+++ assets/js/components/login-component/styles.less
@@ -1,0 +1,12 @@
+login-component {
+  #login-modal {
+
+    &.visible {
+      display: block;
+    }
+
+    label {
+      color: red !important;
+    }
+  }
+}
